+++ lib/src/config/config_import.dart
@@ -6,8 +6,10 @@ import 'package:opticore/opticore.dart';
 import '../utils/ui/core_assets.dart';
 
 part '../config/infrastructure/core_setup.dart';
+part '../config/settings/api_response_config.dart';
 part '../config/settings/app_config.dart';
 part '../config/settings/maintenance_config.dart';
 part '../config/settings/network_config.dart';
 part '../config/settings/no_internet_config.dart';
 part '../config/settings/not_found_config.dart';
+part '../config/settings/unauthenticated_config.dart';