+++ lib/src/extensions/extensions_import.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:opticore/opticore.dart';
@@ -6,6 +8,7 @@ part '../extensions/data/context_extension.dart';
 part '../extensions/data/datetime_extension.dart';
 part '../extensions/data/group_extension.dart';
 part '../extensions/data/hex_color_extension.dart';
+part '../extensions/data/let_extension.dart';
 part '../extensions/data/map_extension.dart';
 part '../extensions/data/navigation_extension.dart';
 part '../extensions/data/precache_extension.dart';